# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "媒资管理系统 REST API。成功响应直接返回业务 JSON（camelCase）；失败返回 { err_code, err_msg }。",
        "title": "GIN-MAM API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/asset": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "直接创建媒资（已有 storagePath，非分片上传流程）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "媒资"
                ],
                "summary": "创建媒资",
                "parameters": [
                    {
                        "description": "媒资参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_domain_asset.createRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_domain_asset.AssetVO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    }
                }
            }
        },
...
        "/api/v1/asset/page": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "媒资"
                ],
                "summary": "分页查询媒资",
                "parameters": [
                    {
                        "description": "分页参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_domain_asset.pageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_domain_asset.PageResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    }
                }
            }
        },
        "/api/v1/asset/upload/chunk": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "tags": [
                    "媒资上传"
                ],
                "summary": "上传分片",
                "parameters": [
                    {
                        "type": "string",
                        "description": "上传会话 ID",
                        "name": "uploadId",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "分片序号（从 0 开始）",
                        "name": "chunkIndex",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "分片二进制",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    }
                }
            }
        },
        "/api/v1/asset/upload/complete": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "合并分片、上传对象存储并创建媒资记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "媒资上传"
                ],
                "summary": "完成分片上传",
                "parameters": [
                    {
                        "description": "完成参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_domain_asset.uploadCompleteRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_domain_asset.AssetVO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    }
                }
            }
        },
        "/api/v1/asset/upload/init": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "媒资上传"
                ],
                "summary": "初始化分片上传",
                "parameters": [
                    {
                        "description": "文件信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_domain_asset.uploadInitRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_domain_asset.UploadInitResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    }
                }
            }
        },
        "/api/v1/asset/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "媒资"
                ],
                "summary": "获取媒资详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "媒资 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_domain_asset.AssetVO"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "媒资"
                ],
                "summary": "更新媒资",
                "parameters": [
                    {
                        "type": "string",
                        "description": "媒资 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_domain_asset.updateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_domain_asset.AssetVO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "软删除媒资（GORM DeletedAt）",
                "tags": [
                    "媒资"
                ],
                "summary": "删除媒资",
                "parameters": [
                    {
                        "type": "string",
                        "description": "媒资 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo": {
            "type": "object",
            "properties": {
                "err_code": {
                    "type": "integer",
                    "example": 40100
                },
                "err_msg": {
                    "type": "string",
                    "example": "未授权"
                }
            }
        },
        "internal_domain_asset.AssetVO": {
            "type": "object",
            "properties": {
                "catalogId": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "previewUrl": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "storagePath": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "internal_domain_asset.PageResult": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_domain_asset.AssetVO"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "internal_domain_asset.UploadInitResult": {
            "type": "object",
            "properties": {
                "chunkSize": {
                    "type": "integer"
                },
                "uploadId": {
                    "type": "string"
                }
            }
        },
        "internal_domain_asset.createRequest": {
            "type": "object",
            "required": [
                "storagePath",
                "title",
                "type"
            ],
            "properties": {
                "catalogId": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "fileSize": {
                    "type": "integer"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "mimeType": {
                    "type": "string"
                },
                "storagePath": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "internal_domain_asset.pageRequest": {
            "type": "object",
            "properties": {
                "keyword": {
                    "type": "string",
                    "example": ""
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "pageSize": {
                    "type": "integer",
                    "example": 20
                }
            }
        },
        "internal_domain_asset.updateRequest": {
            "type": "object",
            "properties": {
                "catalogId": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "status": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "internal_domain_asset.uploadCompleteRequest": {
            "type": "object",
            "required": [
                "assetTitle",
                "type",
                "uploadId"
            ],
            "properties": {
                "assetTitle": {
                    "type": "string"
                },
                "catalogId": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "uploadId": {
                    "type": "string"
                }
            }
        },
        "internal_domain_asset.uploadInitRequest": {
            "type": "object",
            "required": [
                "fileName",
                "fileSize"
            ],
            "properties": {
                "chunkSize": {
                    "type": "integer"
                },
                "fileName": {
                    "type": "string"
                },
                "fileSize": {
                    "type": "integer"
                },
                "mimeType": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Bearer Token，格式：Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}